corrected bug & moved bw_engine completely in MultiParticleContaine
