+++ Source/Particles/MultiParticleContainer.H
@@ -22,6 +22,37 @@
 // concrete class dervied from WarpXParticlContainer.
 //
 
+//Enable QED effects
+#define WARPX_QED
+
+#ifdef WARPX_QED
+//BW ENGINE
+//#define PXRMP_GPU __host__ __device__
+#define PXRMP_WITH_SI_UNITS
+#include "breit_wheeler_engine.hpp"
+//Alias for the picsar::multi_physics namespace
+namespace pxrmp =  picsar::multi_physics;
+
+//RNG wrapper BW engine
+class amrex_rng_wrapper
+{
+public:
+    //Get rnd number uniformly distributed in [a,b)
+    amrex::Real unf(amrex::Real a, amrex::Real b)
+    {
+        return (b-a)*amrex::Random() + a;
+    }
+
+    //Get rnd number with exponential distribution
+    amrex::Real exp(amrex::Real l)
+    {
+        amrex::Real zero_plus_to_one = 1.0 - unf(0.0, 1.0);
+        return -log(zero_plus_to_one)/l;
+    }
+};
+//___________________________________________
+#endif
+
 class MultiParticleContainer
 {
 
@@ -198,9 +229,9 @@ public:
     PhysicalParticleContainer& GetPCtmp () { return *pc_tmp; }
 
     //bw_engine is a public member of the MultiParticleContainer object
-#ifdef WARPX_QED
+//#ifdef WARPX_QED
     pxrmp::breit_wheeler_engine<amrex::Real, amrex_rng_wrapper> bw_engine;
-#endif
+//#endif
 
 protected:
